Created Graphics Class to handle graphical elements
